webapp-clean/requirements.txt
Flask==3.0.0
pyodbc==5.0.1
azure-identity==1.15.0
azure-core==1.29.5
Werkzeug==3.0.1
gunicorn==21.2.0
